+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Build MATLAB toolbox in actions.
+
 ### Changed
 
-- Update to matlab-igraph 0.1.9
+- Update to matlab-igraph 0.1.9.
 - Move argument parsing to MATLAB.
 - Move from GNUMake to CMake.
 - Reorganize toolbox to remove C source code from distributed package.